+++ Tests/SentryTests/SentryCrash/SentryStacktraceBuilderTests.swift
@@ -5,21 +5,21 @@ class SentryStacktraceBuilderTests: XCTestCase {
     
     private class Fixture {
         func getSut() -> SentryStacktraceBuilder {
-            SentryStacktraceBuilder()
+            SentryStacktraceBuilder(sentryFrameRemover: SentryFrameRemover())
         }
     }
     
     private let fixture = Fixture()
     
     func testEnoughFrames() {
-        let actual = fixture.getSut().buildStacktraceForCurrentThread(framesToSkip: 0)
+        let actual = fixture.getSut().buildStacktraceForCurrentThread()
         
         // The stacktrace has usually more than 40 frames. Feel free to change the number if the tests are failing
         XCTAssertTrue(30 < actual.frames.count, "Not enough stacktrace frames.")
     }
     
     func testFramesAreFilled() {
-        let actual = fixture.getSut().buildStacktraceForCurrentThread(framesToSkip: 0)
+        let actual = fixture.getSut().buildStacktraceForCurrentThread()
         
         // We don't know the actual values of the frames so we can't write
         // deterministic tests here. Therefore we just make sure they are
@@ -33,7 +33,7 @@ class SentryStacktraceBuilderTests: XCTestCase {
     }
     
     func testFramesDontContainBuilderFunction() {
-        let actual = fixture.getSut().buildStacktraceForCurrentThread(framesToSkip: 0)
+        let actual = fixture.getSut().buildStacktraceForCurrentThread()
         
         let result = actual.frames.contains { frame in
             return frame.function?.contains("buildStacktraceForCurrentThread") ?? false
@@ -43,34 +43,10 @@ class SentryStacktraceBuilderTests: XCTestCase {
     }
     
     func testFramesOrder() {
-        let actual = fixture.getSut().buildStacktraceForCurrentThread(framesToSkip: 0)
-        let lastFrame = actual.frames.last
-        let areFramesOrderedCorrect = lastFrame?.function?.contains("testFramesOrder") ?? false
+        let actual = fixture.getSut().buildStacktraceForCurrentThread()
+        let firstFrame = actual.frames.first
+        let areFramesOrderedCorrect = firstFrame?.function?.contains("start") ?? false
         
         XCTAssertTrue(areFramesOrderedCorrect, "The frames must be ordered from caller to callee, or oldest to youngest.")
-    }
-    
-    func testSkippingFrames() {
-        // This function should be removed from the stacktrace
-        func wrapperFunc() -> Stacktrace {
-            fixture.getSut().buildStacktraceForCurrentThread(framesToSkip: 1)
-        }
-        let actual = wrapperFunc()
-        let result = actual.frames.contains { frame in
-            return frame.function?.contains("wrapperFunc") ?? false
-        }
-        
-        XCTAssertFalse(result, "The stacktrace should not contain the wrapperFunc.")
-        
-         let noSkipping = fixture.getSut().buildStacktraceForCurrentThread(framesToSkip: 0)
-        // The count must be equal because the stacktrace of actual has one more
-        // function.
-        XCTAssertEqual(noSkipping.frames.count, actual.frames.count)
-    }
-    
-    func testSkippingAllFrames() {
-        let actual = fixture.getSut().buildStacktraceForCurrentThread(framesToSkip: 1_000)
-        
-        XCTAssertEqual(0, actual.frames.count)
     }
 }