+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## unreleased
 
+- feat: Replace passing nullable Scope with overloads #743 !Breaking
+- feat: Remove SDK frames when attaching stacktrace #739
+
+## 6.0.0-beta.2
+
+Breaking changes:
+
+- feat: Remove SentryOptions.enabled #736
+- fix: Public Headers #735
+
+Fix:
 - fix: Setting environment for Sessions #734
 
 ## 6.0.0-beta.1